Issue #17 - Move storage and local store from utils to separated files
