+++ src/js/migration.js
@@ -4,8 +4,6 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import _ from "lodash";
 
 import {
-  Storage,
-  LocalStore,
   ConfigFile,
 } from "./utils";
 import {
@@ -13,6 +11,8 @@ import {
   StoreKey,
 } from "./shared/constants";
 import { isDate } from "./shared/misc";
+import LocalStore from "./shared/local_store";
+import Storage from "./shared/storage";
 
 /**
  * This script is run when updating the extension from version 4.x.x to version 5.0.0 (or higher) and
