+++ src/js/postHandler.js
@@ -4,12 +4,8 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import _ from "lodash";
 import Mustache from "mustache";
 
-import {
-  LocalStore
-} from "./utils";
-import {
-  StoreKey,
-} from "./shared/constants";
+import LocalStore from "./shared/local_store";
+import { StoreKey } from "./shared/constants";
 
 // Wait for the page to be loaded to execute the initialization function.
 document.addEventListener("DOMContentLoaded", function() {
